scripts/requirements.txt
openpyxl>=3.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pypinyin>=0.50.0
python-dotenv>=1.0.0
